Ruter alle journalposter til k9sak Etter diskusjon på standup kom vi fram til att allt i punsj rutes direkt til k9sak, eventuelle infotrygd-koblinger fanges opp i k9sak. Flyttet KopierJournalpost route in i journalpostroutes-klasse
