+++ app/src/main/kotlin/no/nav/k9punsj/journalpost/dto/KopierJournalpostDto.kt
@@ -1,0 +1,14 @@
+package no.nav.k9punsj.journalpost.dto
+
+data class KopierJournalpostDto(
+    val fra: String,
+    val til: String,
+    val barn: String?,
+    val annenPart: String?
+) {
+    init {
+        require(barn != null || annenPart != null) {
+            "Må sette minst en av barn og annenPart"
+        }
+    }
+}
